discovery: cohesion III
Editing disabled services rules is rulesets business.
